Fix default Config marshaling weirdly
Basically the config took a string input on the file, but the internal had a different type, which meant when you "wrote" the config file, it would write the *entire* byte array of that other type.

I modified the `text(str)` function to instead be an global function to be used in the exactly *two* places that needed it. rather than on the global config struct